Implemented idea for turning using the research paper

+++ robot/rover/MobilePlatform/includes/Rover.h
@@ -104,7 +104,7 @@ namespace Rover {
 
     void closeLoop();
 
-    void moveRover(const uint8_t& throttle_direction, const uint8_t & throttle, const uint8_t & steer_direction,const uint8_t & steering);
+    void moveRover(const uint8_t & throttle_direction, const float & throttle, const uint8_t & steer_direction,const float & steering);
 
     void stopMotors();
